Job-Agent-AI-main/requirements.txt
fastapi
uvicorn
python-multipart
python-jobspy
google-generativeai
fpdf2
pypdf
selenium
webdriver-manager
pandas
python-dotenv
